✨ volumeColor & buttonColor properties now change css
🧪 Added unit test for buttonColor

🔥 Removed jest badges. This is done through codecov now.
