+++ packages/custoplayer/package.json
@@ -45,7 +45,6 @@
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-sonarjs": "^0.18.0",
     "jest": "^29.4.3",
-    "jest-coverage-badges": "^1.1.2",
     "jest-environment-jsdom": "^29.4.3",
     "json": "11.0.0",
     "prettier": "^2.8.4",
